django_api/requirements.txt
django==4.2.11
python-dotenv==1.0.1
django-split-settings==1.3.1
psycopg2-binary==2.9.9
pytest==8.2.1
djangorestframework==3.15.1
uwsgi==2.0.26